xcp/cpiofile.py: Fix is_cpiofile() and opening LZMA with Py3
Fix the "FIXME: not compatible with python3 API":

- Fix xzopen() to use the Python3 API when on Python3.

- Fix xcp.cpiofile.is_cpiofile() for LZMA files to return true
  and False for files whicha are not supported:

When xcp.cpiofile.is_cpiofile() is called with a file which is not
an valid cpio/cpio.gz/cpio.bz2 file, cpiofile would try to
call "xzopen ", which does not exist.

After fixing this typo, if the file is also valid XZ/LZMA,
lzma.error(from Py2) or lzma.LZMAError (from Py3) is raised,
not IOError as assumed in the xzopen function.

If lzma.error(from Py2) or lzma.LZMAError not catched, these
are also not catched bt xcp.cpiofile.is_cpiofile() which only
catches CpioError, to return False to indicate No, not a cpio file.

Catch the respective lzma exceptions and raise them as instances
CpioError for xcp.cpiofile.is_cpiofile() to work.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>