xcp/cpiofile.py: Add type comments to fix static type checking
This adds type comments to fix static type checking and a cast
with the goal of being able to to the pyright type checker which is
also used by the language backend of the Python extension of vscode.

It can be also used standalone for CI/CD, even in GitHub actions.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>